Create modules/models_settings.py
